pyproject.toml
[project]
name = "agentapi"
version = "0.1.0"
description = ""
authors = [
    {name = "ldyislucky",email = "1278950466@qq.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain (>=0.3.22,<0.4.0)",
    "langchain-deepseek (>=0.1.3,<0.2.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)",
    "langgraph (>=0.3.24,<0.4.0)",
    "fastapi (>=0.115.9,<0.116.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "redis (>=5.2.1,<6.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "docx2txt (>=0.9,<0.10)",
    "sentence-transformers (>=4.0.2,<5.0.0)",
    "chromadb (>=1.0.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pillow (>=11.2.1,<12.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "dbutils (>=3.1.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "agentapi", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.poetry.dependencies]
python = ">=3.12, <4.0"  # 添加明确的 Python 版本上限